Rename no-useless-template-literals to no-useless-template-expression
