+++ packages/eslint-plugin/docs/rules/no-useless-template-literals.mdx
@@ -1,5 +1,5 @@
 ---
-description: 'Disallow unnecessary template literals.'
+description: 'Disallow unnecessary template expressions.'
 ---
 
 import Tabs from '@theme/Tabs';
@@ -9,53 +9,15 @@ import TabItem from '@theme/TabItem';
 >
 > See **https://typescript-eslint.io/rules/no-useless-template-literals** for documentation.
 
-This rule reports template literals that can be simplified to a normal string literal.
+This rule reports template literals that contain substitution expressions (also variously referred to as embedded expressions or string interpolations) that are unnecessary and can be simplified.
 
-## Examples
+:::warning
+This rule is being renamed to [`no-useless-template-expression`](./no-useless-template-expression.mdx).
+After the creation of this rule, it was realized that the name `no-useless-template-literals` could be misleading, seeing as this rule only targets template literals with substitution expressions.
+In particular, it does _not_ aim to flag useless template literals that look like `` `this` `` and could be simplified to `"this"`.
+If you are looking for such a rule, you can configure the [`quotes`](./quotes.mdx) rule to do this.
 
-<Tabs>
-<TabItem value="❌ Incorrect">
+The current name, `no-useless-template-literals`, will be removed in a future major version of typescript-eslint.
+:::
 
-```ts
-const ab1 = `${'a'}${'b'}`;
-const ab2 = `a${'b'}`;
-
-const stringWithNumber = `${'1 + 1 = '}${2}`;
-
-const stringWithBoolean = `${'true is '}${true}`;
-
-const text = 'a';
-const wrappedText = `${text}`;
-
-declare const intersectionWithString: string & { _brand: 'test-brand' };
-const wrappedIntersection = `${intersectionWithString}`;
-```
-
-</TabItem>
-<TabItem value="✅ Correct">
-
-```ts
-const ab1 = 'ab';
-const ab2 = 'ab';
-
-const stringWithNumber = `1 + 1 = 2`;
-
-const stringWithBoolean = `true is true`;
-
-const text = 'a';
-const wrappedText = text;
-
-declare const intersectionWithString: string & { _brand: 'test-brand' };
-const wrappedIntersection = intersectionWithString;
-```
-
-</TabItem>
-</Tabs>
-
-## When Not To Use It
-
-When you want to allow string expressions inside template literals.
-
-## Related To
-
-- [`restrict-template-expressions`](./restrict-template-expressions.mdx)
+{/* Intentionally Omitted: When Not To Use It */}
