+++ packages/eslint-plugin/src/configs/strict-type-checked.ts
@@ -60,7 +60,7 @@ export = {
     '@typescript-eslint/no-unused-vars': 'error',
     'no-useless-constructor': 'off',
     '@typescript-eslint/no-useless-constructor': 'error',
-    '@typescript-eslint/no-useless-template-literals': 'error',
+    '@typescript-eslint/no-useless-template-expression': 'error',
     '@typescript-eslint/no-var-requires': 'error',
     'no-throw-literal': 'off',
     '@typescript-eslint/only-throw-error': 'error',
